chore: Splitted data.ts file into separate files
