+++ src/app/(auth)/(routes)/login/[[...login]]/page.tsx
@@ -7,7 +7,7 @@ import { useForm } from "react-hook-form"
 import sanitizeHtml from "sanitize-html"
 import { isThirdPartyAuthenticationEnabled } from "@/lib/utils"
 import { LoginFormValuesType, loginSchema } from "@/lib/validation/auth"
-import { signUpUrl } from "@/data"
+import { signUpUrl } from "@/data/authData"
 import { AppLogo } from "@/components/main/app-logo"
 import { Form } from "@/components/ui/form"
 import { CustomFormLink } from "@/components/form/custom-form-link"
