+++ src/app/(auth)/(routes)/register/[[...register]]/page.tsx
@@ -5,7 +5,7 @@ import { useForm, useWatch } from "react-hook-form"
 import { zodResolver } from "@hookform/resolvers/zod"
 import { isThirdPartyAuthenticationEnabled } from "@/lib/utils"
 import { RegisterFormValuesType, registerSchema } from "@/lib/validation/auth"
-import { signInUrl } from "@/data"
+import { signInUrl } from "@/data/authData"
 import { AppLogo } from "@/components/main/app-logo"
 import { Form } from "@/components/ui/form"
 import { CustomFormInput } from "@/components/form/custom-form-input"
