+++ src/app/(public)/nitro/page.tsx
@@ -7,7 +7,8 @@ import { useTheme } from "next-themes"
 import { Check, Plus, X } from "lucide-react"
 import sanitizeHtml from "sanitize-html"
 import { cn } from "@/lib/utils"
-import { nitroBannerList, nitroFAQCategoryList, nitroFAQList, nitroMainPerkList, nitroPlanFeatureList, nitroPlanList, nitroSubPerkList, placeholderImageUrl } from "@/data"
+import { placeholderImageUrl } from "@/data/globalData"
+import { nitroBannerList, nitroFAQCategoryList, nitroFAQList, nitroMainPerkList, nitroPlanFeatureList, nitroPlanList, nitroSubPerkList } from "@/data/nitroPageData"
 import { PageLayout } from "@/components/main/page-layout"
 import CustomVideo from "@/components/main/custom-video"
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs"
