+++ src/components/main/footer.tsx
@@ -6,7 +6,8 @@ import Image from "next/image"
 import { useAuth } from "@clerk/nextjs"
 import { cn } from "@/lib/utils"
 import { FooterProps } from "@/types"
-import { footerMenuItems, placeholderImageUrl, afterSignInUrl, signUpUrl, socialLinks } from "@/data"
+import { footerMenuItems, placeholderImageUrl, socialLinks } from "@/data/globalData"
+import { afterSignInUrl, signUpUrl } from "@/data/authData"
 import { Accordion, AccordionContent, AccordionItem, AccordionTrigger } from "@/components/ui/accordion"
 import { LanguageDropdown } from "@/components/language-dropdown"
 import { AppLogo } from "@/components/main/app-logo"
