+++ src/components/main/header.tsx
@@ -6,7 +6,7 @@ import Link from "next/link"
 import { useAuth } from "@clerk/nextjs"
 import { cn } from "@/lib/utils"
 import { HeaderProps } from "@/types"
-import { afterSignInUrl, signInUrl } from "@/data"
+import { afterSignInUrl, signInUrl } from "@/data/authData"
 import { AppLogo } from "@/components/main/app-logo"
 import { NavigationMenuCustom } from "@/components/main/navigation-menu/navigation-menu-custom"
 import NavigationBurgerMenu from "@/components/main/navigation-menu/navigation-burger-menu"
