+++ src/components/main/navigation-menu/navigation-menu-item-custom.tsx
@@ -3,7 +3,7 @@ import Image from "next/image"
 import { cn } from "@/lib/utils"
 import { ArrowUpRight } from "lucide-react"
 import { NavigationMenuCustomProps, NavigationMenuItemContentProps } from "@/types"
-import { placeholderImageUrl } from "@/data"
+import { placeholderImageUrl } from "@/data/globalData"
 import { NavigationMenuContent, NavigationMenuItem, NavigationMenuLink, NavigationMenuTrigger, navigationMenuTriggerStyle } from "@/components/ui/navigation-menu"
 import NavigationMenuListItemTitle from "@/components/main/navigation-menu/navigation-menu-list-item-title"
 import NavigationMenuListItem from "@/components/main/navigation-menu/navigation-menu-list-item"
